refactor(decomp): extract DecompVU / init_decomp_vu / site_out into components.py
The decomposition representation lived in `llama8b.py`, so every target — including the
positionless TMS/ResidMLP toys — imported `DecompVU` / `init_decomp_vu` from a file named
`llama8b`, and `_site_out` (the SPEC §4.1 decomposed linear `((x@V)*m)@U + (x@Δ)*d`) was
byte-identically copy-pasted across `llama8b`, `llama_simple_mlp`, and both toys (a fourth
importer pulled the private `_site_out`).

Move the three domain-neutral primitives into a new `param_decomp/components.py`
(`_site_out` -> public `site_out`, single definition), and repoint all 24 importers. No
target imports the decomposition representation from `llama8b` anymore (AST-verified), and
the §4.1 primitive is one copy.

Verbatim move — values unchanged: `make type` clean, 107 tests pass (the stacked-parity
`out::site_input` goldens confirm bit-identity).
