+++ param_decomp/experiments/llama8b_real.py
@@ -44,6 +44,7 @@ from param_decomp.ci_fn import (
     ChunkwiseTransformerCIArch,
     build_ci_fn,
 )
+from param_decomp.components import init_decomp_vu
 from param_decomp.configs import (
     AdamPGDConfig,
     ChunkwiseSubsetReconLossConfig,
@@ -58,7 +59,6 @@ from param_decomp.llama8b import (
     FrozenAttn,
     SuffixLayer,
     Target,
-    init_decomp_vu,
     llama31_8b_config,
     llama_decomposed_lm,
     llama_site_specs,
