+++ param_decomp/experiments/mem_probe.py
@@ -38,6 +38,7 @@ from param_decomp.ci_fn import (
     ChunkwiseTransformerCIArch,
     build_ci_fn,
 )
+from param_decomp.components import init_decomp_vu
 from param_decomp.configs import (
     AdamPGDConfig,
     ChunkwiseSubsetReconLossConfig,
@@ -51,7 +52,6 @@ from param_decomp.experiments.llama8b_real import (
     _random_target,
 )
 from param_decomp.llama8b import (
-    init_decomp_vu,
     llama31_8b_config,
     llama_decomposed_lm,
     llama_site_specs,
