+++ param_decomp/llama8b_sharding.py
@@ -34,8 +34,9 @@ from jaxtyping import Array, PRNGKeyArray
 
 from param_decomp.adversary import init_persistent_sources
 from param_decomp.ci_fn import CIFn, CIFnArch, build_ci_fn
+from param_decomp.components import DecompVU, init_decomp_vu
 from param_decomp.configs import BSCScope, PersistentPGDSourceScope, SCScope
-from param_decomp.llama8b import DecompVU, Target, init_decomp_vu
+from param_decomp.llama8b import Target
 from param_decomp.lm import SiteSpec
 from param_decomp.log import logger
 from param_decomp.sharding import dp_mesh
