+++ param_decomp/tests/test_checkpoint.py
@@ -26,6 +26,7 @@ from param_decomp.ci_fn import (
     ChunkwiseTransformerCIArch,
     build_ci_fn,
 )
+from param_decomp.components import init_decomp_vu
 from param_decomp.configs import (
     AdamPGDConfig,
     ChunkwiseSubsetReconLossConfig,
@@ -36,7 +37,6 @@ from param_decomp.configs import (
     UniformKSubsetRoutingConfig,
 )
 from param_decomp.llama8b import (
-    init_decomp_vu,
     llama_decomposed_lm,
     llama_site_specs,
     mlp_family_site_cs,
