+++ param_decomp/tests/test_source_grad_mean.py
@@ -39,8 +39,8 @@ from jax.sharding import PartitionSpec as P
 
 from param_decomp.adversary import init_persistent_sources, source_masks
 from param_decomp.ci_fn import Chunk, ChunkwiseTransformerCIArch, build_ci_fn
+from param_decomp.components import init_decomp_vu
 from param_decomp.llama8b import (
-    init_decomp_vu,
     llama_decomposed_lm,
     llama_site_specs,
     mlp_family_site_cs,
