+++ param_decomp_lab/experiments/resid_mlp/test_resid_mlp.py
@@ -22,13 +22,13 @@ from param_decomp.ci_fn import (
     init_global_mlp_ci_fn,
     init_layerwise_mlp_ci_fn,
 )
+from param_decomp.components import DecompVU, init_decomp_vu
 from param_decomp.configs import (
     FaithfulnessLossConfig,
     ImportanceMinimalityLossConfig,
     StochasticReconLayerwiseLossConfig,
     StochasticReconLossConfig,
 )
-from param_decomp.llama8b import DecompVU, init_decomp_vu
 from param_decomp.lm import DecomposedModel, SiteC, SiteSpec
 from param_decomp.recon import build_recon_terms
 from param_decomp.train import TrainState, make_faith_warmup_step, make_train_step
